Start work on EventQueue and InputThread
This is where user interaction is starting to take shape. The plan is to
have auxilliary threads send Event messages to the main thread on a
central EventQueue. This simplify's the main thread's job to handling
whatever messages it receives via the appropriate control actions (such
as pausing a queue when we get a mpl_PAUSE message). This pause message
could come from the user pressing a pause button, or the AudioBackend
running out of tracks to play.