Fix AVPacket resource leak
The leak was only occuring with tracks that have embedded album art,
as that caused a packet with the art to be read and discarded without
being unref'd, causing the underlying buffers to leak. The cause was
assuming that av_read_frame would unref packets automatically, similar
to how avcodec_receive_frame auto-unrefs frames. It does not.