Mostly implement FAST backend
There are some things we need to implement for FAST before its backend
implementation can be truly completed. What'primarily needed is (1) a loop
structure that can be locked/unlocked (this lock is automatically
acquired for callbacks) and (2) a begin_write method that will give us
the max write size that can be passed to FastStream_write.